# service/room/api/room/v1/room_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: room/v1/room.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationRoomServiceCreateRoom = "/room.v1.RoomService/CreateRoom"
const OperationRoomServiceDeleteRoom = "/room.v1.RoomService/DeleteRoom"
const OperationRoomServiceGetRoom = "/room.v1.RoomService/GetRoom"
const OperationRoomServiceJoinRoom = "/room.v1.RoomService/JoinRoom"
const OperationRoomServiceKickMember = "/room.v1.RoomService/KickMember"
const OperationRoomServiceLeaveRoom = "/room.v1.RoomService/LeaveRoom"
const OperationRoomServiceListMembers = "/room.v1.RoomService/ListMembers"
const OperationRoomServiceListUserRooms = "/room.v1.RoomService/ListUserRooms"
const OperationRoomServiceMuteMember = "/room.v1.RoomService/MuteMember"
const OperationRoomServiceUpdateMemberRole = "/room.v1.RoomService/UpdateMemberRole"
const OperationRoomServiceUpdateRoom = "/room.v1.RoomService/UpdateRoom"

type RoomServiceHTTPServer interface {
	// CreateRoom 创建房间
	CreateRoom(context.Context, *CreateRoomRequest) (*CreateRoomReply, error)
	// DeleteRoom 删除房间
	DeleteRoom(context.Context, *DeleteRoomRequest) (*DeleteRoomReply, error)
	// GetRoom 获取房间信息
	GetRoom(context.Context, *GetRoomRequest) (*GetRoomReply, error)
	// JoinRoom 加入房间
	JoinRoom(context.Context, *JoinRoomRequest) (*JoinRoomReply, error)
	// KickMember 踢出成员
	KickMember(context.Context, *KickMemberRequest) (*KickMemberReply, error)
	// LeaveRoom 退出房间
	LeaveRoom(context.Context, *LeaveRoomRequest) (*LeaveRoomReply, error)
	// ListMembers 获取房间成员列表
	ListMembers(context.Context, *ListMembersRequest) (*ListMembersReply, error)
	// ListUserRooms 获取用户加入的房间列表
	ListUserRooms(context.Context, *ListUserRoomsRequest) (*ListUserRoomsReply, error)
	// MuteMember 禁言/解禁成员
	MuteMember(context.Context, *MuteMemberRequest) (*MuteMemberReply, error)
	// UpdateMemberRole 更新成员角色
	UpdateMemberRole(context.Context, *UpdateMemberRoleRequest) (*UpdateMemberRoleReply, error)
	// UpdateRoom 更新房间信息
	UpdateRoom(context.Context, *UpdateRoomRequest) (*UpdateRoomReply, error)
}

func RegisterRoomServiceHTTPServer(s *http.Server, srv RoomServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/rooms", _RoomService_CreateRoom0_HTTP_Handler(srv))
	r.GET("/v1/rooms/{id}", _RoomService_GetRoom0_HTTP_Handler(srv))
	r.PUT("/v1/rooms/{id}", _RoomService_UpdateRoom0_HTTP_Handler(srv))
	r.DELETE("/v1/rooms/{id}", _RoomService_DeleteRoom0_HTTP_Handler(srv))
	r.POST("/v1/rooms/{room_id}/join", _RoomService_JoinRoom0_HTTP_Handler(srv))
	r.POST("/v1/rooms/{room_id}/leave", _RoomService_LeaveRoom0_HTTP_Handler(srv))
	r.GET("/v1/rooms/{room_id}/members", _RoomService_ListMembers0_HTTP_Handler(srv))
	r.DELETE("/v1/rooms/{room_id}/members/{user_id}", _RoomService_KickMember0_HTTP_Handler(srv))
	r.PUT("/v1/rooms/{room_id}/members/{user_id}/role", _RoomService_UpdateMemberRole0_HTTP_Handler(srv))
	r.PUT("/v1/rooms/{room_id}/members/{user_id}/mute", _RoomService_MuteMember0_HTTP_Handler(srv))
	r.GET("/v1/users/{user_id}/rooms", _RoomService_ListUserRooms0_HTTP_Handler(srv))
}

func _RoomService_CreateRoom0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateRoomRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceCreateRoom)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateRoom(ctx, req.(*CreateRoomRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateRoomReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_GetRoom0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetRoomRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceGetRoom)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetRoom(ctx, req.(*GetRoomRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetRoomReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_UpdateRoom0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateRoomRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceUpdateRoom)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateRoom(ctx, req.(*UpdateRoomRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateRoomReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_DeleteRoom0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteRoomRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceDeleteRoom)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteRoom(ctx, req.(*DeleteRoomRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteRoomReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_JoinRoom0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in JoinRoomRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceJoinRoom)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.JoinRoom(ctx, req.(*JoinRoomRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*JoinRoomReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_LeaveRoom0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LeaveRoomRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceLeaveRoom)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LeaveRoom(ctx, req.(*LeaveRoomRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LeaveRoomReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_ListMembers0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListMembersRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceListMembers)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListMembers(ctx, req.(*ListMembersRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListMembersReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_KickMember0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in KickMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceKickMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.KickMember(ctx, req.(*KickMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*KickMemberReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_UpdateMemberRole0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMemberRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceUpdateMemberRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMemberRole(ctx, req.(*UpdateMemberRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateMemberRoleReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_MuteMember0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in MuteMemberRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceMuteMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.MuteMember(ctx, req.(*MuteMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MuteMemberReply)
		return ctx.Result(200, reply)
	}
}

func _RoomService_ListUserRooms0_HTTP_Handler(srv RoomServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListUserRoomsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoomServiceListUserRooms)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListUserRooms(ctx, req.(*ListUserRoomsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListUserRoomsReply)
		return ctx.Result(200, reply)
	}
}

type RoomServiceHTTPClient interface {
	// CreateRoom 创建房间
	CreateRoom(ctx context.Context, req *CreateRoomRequest, opts ...http.CallOption) (rsp *CreateRoomReply, err error)
	// DeleteRoom 删除房间
	DeleteRoom(ctx context.Context, req *DeleteRoomRequest, opts ...http.CallOption) (rsp *DeleteRoomReply, err error)
	// GetRoom 获取房间信息
	GetRoom(ctx context.Context, req *GetRoomRequest, opts ...http.CallOption) (rsp *GetRoomReply, err error)
	// JoinRoom 加入房间
	JoinRoom(ctx context.Context, req *JoinRoomRequest, opts ...http.CallOption) (rsp *JoinRoomReply, err error)
	// KickMember 踢出成员
	KickMember(ctx context.Context, req *KickMemberRequest, opts ...http.CallOption) (rsp *KickMemberReply, err error)
	// LeaveRoom 退出房间
	LeaveRoom(ctx context.Context, req *LeaveRoomRequest, opts ...http.CallOption) (rsp *LeaveRoomReply, err error)
	// ListMembers 获取房间成员列表
	ListMembers(ctx context.Context, req *ListMembersRequest, opts ...http.CallOption) (rsp *ListMembersReply, err error)
	// ListUserRooms 获取用户加入的房间列表
	ListUserRooms(ctx context.Context, req *ListUserRoomsRequest, opts ...http.CallOption) (rsp *ListUserRoomsReply, err error)
	// MuteMember 禁言/解禁成员
	MuteMember(ctx context.Context, req *MuteMemberRequest, opts ...http.CallOption) (rsp *MuteMemberReply, err error)
	// UpdateMemberRole 更新成员角色
	UpdateMemberRole(ctx context.Context, req *UpdateMemberRoleRequest, opts ...http.CallOption) (rsp *UpdateMemberRoleReply, err error)
	// UpdateRoom 更新房间信息
	UpdateRoom(ctx context.Context, req *UpdateRoomRequest, opts ...http.CallOption) (rsp *UpdateRoomReply, err error)
}

type RoomServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewRoomServiceHTTPClient(client *http.Client) RoomServiceHTTPClient {
	return &RoomServiceHTTPClientImpl{client}
}

// CreateRoom 创建房间
func (c *RoomServiceHTTPClientImpl) CreateRoom(ctx context.Context, in *CreateRoomRequest, opts ...http.CallOption) (*CreateRoomReply, error) {
	var out CreateRoomReply
	pattern := "/v1/rooms"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoomServiceCreateRoom))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteRoom 删除房间
func (c *RoomServiceHTTPClientImpl) DeleteRoom(ctx context.Context, in *DeleteRoomRequest, opts ...http.CallOption) (*DeleteRoomReply, error) {
	var out DeleteRoomReply
	pattern := "/v1/rooms/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoomServiceDeleteRoom))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetRoom 获取房间信息
func (c *RoomServiceHTTPClientImpl) GetRoom(ctx context.Context, in *GetRoomRequest, opts ...http.CallOption) (*GetRoomReply, error) {
	var out GetRoomReply
	pattern := "/v1/rooms/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoomServiceGetRoom))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// JoinRoom 加入房间
func (c *RoomServiceHTTPClientImpl) JoinRoom(ctx context.Context, in *JoinRoomRequest, opts ...http.CallOption) (*JoinRoomReply, error) {
	var out JoinRoomReply
	pattern := "/v1/rooms/{room_id}/join"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoomServiceJoinRoom))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// KickMember 踢出成员
func (c *RoomServiceHTTPClientImpl) KickMember(ctx context.Context, in *KickMemberRequest, opts ...http.CallOption) (*KickMemberReply, error) {
	var out KickMemberReply
	pattern := "/v1/rooms/{room_id}/members/{user_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoomServiceKickMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// LeaveRoom 退出房间
func (c *RoomServiceHTTPClientImpl) LeaveRoom(ctx context.Context, in *LeaveRoomRequest, opts ...http.CallOption) (*LeaveRoomReply, error) {
	var out LeaveRoomReply
	pattern := "/v1/rooms/{room_id}/leave"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoomServiceLeaveRoom))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListMembers 获取房间成员列表
func (c *RoomServiceHTTPClientImpl) ListMembers(ctx context.Context, in *ListMembersRequest, opts ...http.CallOption) (*ListMembersReply, error) {
	var out ListMembersReply
	pattern := "/v1/rooms/{room_id}/members"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoomServiceListMembers))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListUserRooms 获取用户加入的房间列表
func (c *RoomServiceHTTPClientImpl) ListUserRooms(ctx context.Context, in *ListUserRoomsRequest, opts ...http.CallOption) (*ListUserRoomsReply, error) {
	var out ListUserRoomsReply
	pattern := "/v1/users/{user_id}/rooms"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoomServiceListUserRooms))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// MuteMember 禁言/解禁成员
func (c *RoomServiceHTTPClientImpl) MuteMember(ctx context.Context, in *MuteMemberRequest, opts ...http.CallOption) (*MuteMemberReply, error) {
	var out MuteMemberReply
	pattern := "/v1/rooms/{room_id}/members/{user_id}/mute"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoomServiceMuteMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateMemberRole 更新成员角色
func (c *RoomServiceHTTPClientImpl) UpdateMemberRole(ctx context.Context, in *UpdateMemberRoleRequest, opts ...http.CallOption) (*UpdateMemberRoleReply, error) {
	var out UpdateMemberRoleReply
	pattern := "/v1/rooms/{room_id}/members/{user_id}/role"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoomServiceUpdateMemberRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateRoom 更新房间信息
func (c *RoomServiceHTTPClientImpl) UpdateRoom(ctx context.Context, in *UpdateRoomRequest, opts ...http.CallOption) (*UpdateRoomReply, error) {
	var out UpdateRoomReply
	pattern := "/v1/rooms/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoomServiceUpdateRoom))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}